requirements.txt
streamlit
pdfplumber
requests
openai
weaviate-client
python-dotenv
numpy
